Rename: 로그인, 시간표 패키지 정리
