+++ app/src/main/java/com/example/relay/login/models/LogInLocalRes.kt
@@ -1,8 +1,8 @@
-package com.example.relay.login.data
+package com.example.relay.login.models
 
 import com.example.relay.BaseResponse
 import com.google.gson.annotations.SerializedName
 
-data class SignUpLocalRes(
+data class LogInLocalRes(
     @SerializedName("result") val result: BaselResult
 ) : BaseResponse()
