+++ app/src/main/java/com/example/relay/login/models/SignUpLocalReq.kt
@@ -1,8 +1,9 @@
-package com.example.relay.login.data
+package com.example.relay.login.models
 
 import com.google.gson.annotations.SerializedName
 
-data class LogInLocalReq(
+data class SignUpLocalReq(
+    @SerializedName("name") val name: String,
     @SerializedName("email") val email: String,
     @SerializedName("pwd") val pwd: String
 )
