+++ app/src/main/java/com/example/relay/login/models/SignUpLocalRes.kt
@@ -1,8 +1,8 @@
-package com.example.relay.login.data
+package com.example.relay.login.models
 
 import com.example.relay.BaseResponse
 import com.google.gson.annotations.SerializedName
 
-data class LogInLocalRes(
+data class SignUpLocalRes(
     @SerializedName("result") val result: BaselResult
 ) : BaseResponse()
