+++ app/src/main/java/com/example/relay/login/models/UserProfileRes.kt
@@ -1,4 +1,4 @@
-package com.example.relay.login.data
+package com.example.relay.login.models
 
 import com.example.relay.BaseResponse
 import com.google.gson.annotations.SerializedName
