+++ app/src/main/java/com/example/relay/login/view/AuthenticationActivity.kt
@@ -1,9 +1,8 @@
-package com.example.relay.login
+package com.example.relay.login.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.widget.Toast
 
 import com.example.relay.databinding.ActivityAuthenticationBinding
