+++ app/src/main/java/com/example/relay/timetable/service/TimetableInterface.kt
@@ -1,4 +1,4 @@
-package com.example.relay.timetable
+package com.example.relay.timetable.service
 
 import com.example.relay.timetable.models.GroupTimetableRes
 import com.example.relay.timetable.models.MyTimetableRes
