+++ app/src/main/res/layout/activity_authentication.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.AuthenticationActivity">
+    tools:context=".login.view.AuthenticationActivity">
 
     <Button
         android:id="@+id/btn_back"
