+++ app/src/main/res/layout/activity_forget_pw.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.ForgetPwActivity">
+    tools:context=".login.view.ForgetPwActivity">
 
     <TextView
         android:id="@+id/tv_forget"
