+++ app/src/main/res/layout/activity_last_check.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.LastCheckActivity"
+    tools:context=".login.view.LastCheckActivity"
     android:background="@color/background_login">
 
     <ImageButton
