+++ app/src/main/res/layout/activity_login_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.LoginMainActivity"
+    tools:context=".login.view.LoginMainActivity"
     android:background="@color/background_login">
 
     <ImageView
