+++ app/src/main/res/layout/activity_new_pw.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.NewPwActivity">
+    tools:context=".login.view.NewPwActivity">
 
     <TextView
         android:id="@+id/tv_title_new_pw"
