+++ app/src/main/res/layout/activity_privacy_condition.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.PrivacyConditionActivity">
+    tools:context=".login.view.PrivacyConditionActivity">
 
     <TextView
         android:id="@+id/tv_info"
