+++ app/src/main/res/layout/activity_signup.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.SignupActivity"
+    tools:context=".login.view.SignupActivity"
     android:background="@color/background_login">
 
     <TextView
